Moves create_all_accounts_run_and_snapshot_dirs() into accounts-db utils
