+++ ledger-tool/src/ledger_utils.rs
@@ -3,7 +3,9 @@ use {
     clap::{value_t, value_t_or_exit, values_t_or_exit, ArgMatches},
     crossbeam_channel::unbounded,
     log::*,
-    solana_accounts_db::hardened_unpack::open_genesis_config,
+    solana_accounts_db::{
+        hardened_unpack::open_genesis_config, utils::create_all_accounts_run_and_snapshot_dirs,
+    },
     solana_core::{
         accounts_hash_verifier::AccountsHashVerifier, validator::BlockVerificationMethod,
     },
@@ -35,8 +37,8 @@ use {
         snapshot_config::SnapshotConfig,
         snapshot_hash::StartingSnapshotHashes,
         snapshot_utils::{
-            self, clean_orphaned_account_snapshot_dirs, create_all_accounts_run_and_snapshot_dirs,
-            move_and_async_delete_path_contents, SnapshotError,
+            self, clean_orphaned_account_snapshot_dirs, move_and_async_delete_path_contents,
+            SnapshotError,
         },
     },
     solana_sdk::{
@@ -68,7 +70,7 @@ pub(crate) enum LoadAndProcessLedgerError {
     CleanOrphanedAccountSnapshotDirectories(#[source] SnapshotError),
 
     #[error("failed to create all run and snapshot directories: {0}")]
-    CreateAllAccountsRunAndSnapshotDirectories(#[source] SnapshotError),
+    CreateAllAccountsRunAndSnapshotDirectories(#[source] std::io::Error),
 
     #[error("custom accounts path is not supported with seconday blockstore access")]
     CustomAccountsPathUnsupported(#[source] BlockstoreError),
